feat: all types define in seprate file and remove commits
