+++ lib/prismadb.ts
@@ -1,11 +1,11 @@
 import { PrismaClient } from "@prisma/client";
 
-declare global{
-  var prisma: PrismaClient
+declare global {
+  var prisma: PrismaClient;
 }
 
 const db = globalThis.prisma || new PrismaClient();
 
-if(process.env.NODE_ENV === 'production') globalThis.prisma = db;
+if (process.env.NODE_ENV === "production") globalThis.prisma = db;
 
 export default db;
