+++ src/app/(content)/personalInfoContent/PersonalInfoContent.tsx
@@ -1,23 +1,8 @@
 "use client";
+import { PersonalInfoContentProps } from "@/app/types/types";
 import React from "react";
 
-type PersonalInfoContentProps = {
-  userData: {
-    firstName: string;
-    lastName: string;
-    mobileNumber: string;
-    email: string;
-    dateOfBirth: string;
-    maritalStatus: string;
-    gender: string;
-    nationality: string;
-    address: string;
-    city: string;
-    state: string;
-    zipCode: string;
-  };
-  onTabChange: (tabName: string) => void;
-};
+
 
 const InfoSection = ({ label, value }: { label: string; value: string }) => {
   return (
