+++ src/app/types/next-auth.d.ts
@@ -3,22 +3,24 @@ import { JWT } from "next-auth/jwt";
 
 declare module "next-auth" {
   interface User extends DefaultUser {
-    role?: string; 
-    id: string;  
+    id: string;
+    role?: string;
+    avatar?: string | null;
   }
 
   interface Session {
     user: {
       id: string;
       role?: string;
+      avatar?: string | null;
     } & DefaultSession["user"];
   }
 }
 
-
 declare module "next-auth/jwt" {
   interface JWT {
-    role?: string; // Add role to JWT
-    id: string;    // Ensure id is included
+    id: string;
+    role?: string;
+    avatar?: string | null;
   }
 }
